backend/requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
asyncpg==0.30.0
environs==14.1.1
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.1
yarl==1.20.0
python-dotenv==1.1.0
# Админка зависимости
sqladmin==0.21.0
python-multipart==0.0.20
wtforms==3.1.2
werkzeug==3.1.3
itsdangerous==2.1.2
psycopg2-binary==2.9.9